Improved Reset Button / Decimal Zero Fix
Improved the reset button by adding clear function. This clears just
value B if present or clears all if only value A present. Added a
filter to check if value is zero when decimal button pressed so a zero
will be shown in front of the value instead of blank space.